consolidate location of utility functions to /lib
